pyproject.toml
[project]
name = "proxify-cotton"
version = "0.1.0"
description = ""
authors = [
    {name = "carlgro",email = "guerreroaguilar182@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "croniter (>=6.0.0,<7.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "duckdb (>=1.2.0,<2.0.0)",
    "pyarrow (>=19.0.0,<20.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
